lab-tutorial-agent/requirements.txt
fastapi
uvicorn
google-cloud-aiplatform
tenacity
pydantic
google-adk
